requirements.txt
pathlib==1.0.1
Pillow==8.0.1
